+++ go/ndt7-client/sink/measurer.go
@@ -1,4 +1,4 @@
-package protocol
+package sink
 
 import (
 	"encoding/json"
