+++ go/ndt7-client/sink/reader.go
@@ -1,4 +1,4 @@
-package protocol
+package sink
 
 import (
 	"strings"
